refactor(editor): extract selection transforms
Move instance-group rotation/reflection, drafting rotation, single-instance orientation projection, and alignment commands into a selection transform controller. App now wires command routing to the controller instead of implementing transform policy inline.

Validation: pnpm typecheck; pnpm test:local apps/editor/src/features/drafting/drafting-manipulation.test.ts apps/editor/src/features/selection/use-selection-interaction.test.ts apps/editor/src/app/App.test.tsx; pnpm test:e2e:local apps/editor/e2e/manual-editor.spec.ts --grep turns a marquee selection as one body, not three parts in place|flips a marquee selection as one body, not three parts in place; pnpm test:e2e:local apps/editor/e2e/drafting.spec.ts --grep selected arrow rotates via R and shows selection handles; pnpm ci:static; git diff --check

Test-Impact: no-test-change — existing drafting, selection, group-transform, and browser contracts cover unchanged transform behavior
